requirements.txt
newspaper3k==0.2.8
pocket-api==0.1.5
